+++ packages/cli/src/utils/logger.ts
@@ -14,4 +14,7 @@ export const logger = {
   success(...args: unknown[]) {
     console.log(chalk.green(...args));
   },
+  table(rows: any) {
+    console.table(rows);
+  },
 };