+++ packages/database/prisma/migrations/20231013083144_add_next_event_timestamp_to_schedule_source/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "ScheduleSource" ADD COLUMN     "nextEventTimestamp" TIMESTAMP(3);