+++ packages/trigger-sdk/src/apiClient.ts
@@ -31,7 +31,6 @@ import {
   API_VERSIONS,
 } from "@trigger.dev/core";
 
-import fetch, { type RequestInit } from "node-fetch";
 import { z } from "zod";
 
 export type ApiClientOptions = {
@@ -397,6 +396,22 @@ export class ApiClient {
         },
       }
     );
+  }
+
+  async cancelRun(runId: string) {
+    const apiKey = await this.#apiKey();
+
+    this.#logger.debug("Cancelling Run", {
+      runId,
+    });
+
+    return await zodfetch(GetRunSchema, `${this.#apiUrl}/api/v1/runs/${runId}/cancel`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${apiKey}`,
+      },
+    });
   }
 
   async getRunStatuses(runId: string) {