+++ packages/react/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @trigger.dev/react
 
+## 2.2.0
+
+### Patch Changes
+
+- Updated dependencies [975c5f1d]
+- Updated dependencies [50e3d9e4]
+- Updated dependencies [59a94c71]
+  - @trigger.dev/core@2.2.0
+
+## 2.1.9
+
+### Patch Changes
+
+- 9a187f9e: upgrade zod to 3.22.3
+- Updated dependencies [9a187f9e]
+  - @trigger.dev/core@2.1.9
+
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [6a992a19]
+- Updated dependencies [ab9e4a98]
+- Updated dependencies [ab9e4a98]
+  - @trigger.dev/core@2.1.8
+
 ## 2.1.7
 
 ### Patch Changes